+++ test/dummy/test/integration/admin_support_section_test.rb
@@ -35,7 +35,7 @@ class AdminSupportSectionTest < ActionDispatch::IntegrationTest
 
     assert_response :success
     assert_select "body[data-recording-studio-default-layout='true']", count: 1
-    assert_select "html[data-theme='rounded']"
+    assert_select "body[data-theme='rounded']"
     assert_includes response.body, "Help pages"
     assert_includes response.body, "/support/new"
     assert_includes response.body, 'name="search"'